voice_box/requirements.txt
sounddevice
scipy
requests
python-dotenv 